src/main/java/com/winesales/salesmanager/dto/ClienteComprasDTO.java
package com.winesales.salesmanager.dto;

import java.math.BigDecimal;
import java.util.List;

public class ClienteComprasDTO {
    private String nome;
    private String cpf;
    private List<CompraDTO> compras;
    private BigDecimal valorTotal;


    public String getNome() {
        return nome;
    }

    public void setNome(String nome) {
        this.nome = nome;
    }

    public String getCpf() {
        return cpf;
    }

    public void setCpf(String cpf) {
        this.cpf = cpf;
    }

    public List<CompraDTO> getCompras() {
        return compras;
    }

    public void setCompras(List<CompraDTO> compras) {
        this.compras = compras;
    }

    public BigDecimal getValorTotal() {
        return valorTotal;
    }

    public void setValorTotal(BigDecimal valorTotal) {
        this.valorTotal = valorTotal;
    }
}


